Restructure programme
Instead of keeping all the logic in main(), split up the functionality
into multiple structures and methods: There is a Device which allows one
to iterate over variable‐length Chunks, which in turn allow iteration
over Sectors.

Chunks represent the maximum number of contiguous logical sectors which
can be read in a single operation. Every Sector represents a logical
sector in a Chunk.

The overall scrub progress on a volume is represented by a Progress
structure.